flattening file structure, restoring get tweets and sample
