+++ ETL/README.md
@@ -1,0 +1,3 @@
+# ETL README
+
+This directory contains code to load sampled tweets into the database and make all necessary transformations.
