--- src/ETL/README.md
@@ -1,3 +1,0 @@
-# ETL README
-
-This directory contains code to load sampled tweets into the database and make all necessary transformations.
